+++ packages/code-editor/src/CodeEditorCopyButton/CodeEditorCopyButton.tsx
@@ -5,11 +5,12 @@ import { useDarkMode } from '@leafygreen-ui/leafygreen-provider';
 import { keyMap } from '@leafygreen-ui/lib';
 import { palette } from '@leafygreen-ui/palette';
 import { color } from '@leafygreen-ui/tokens';
-import Tooltip, {
+import {
   Align,
   hoverDelay,
   Justify,
   RenderMode,
+  Tooltip,
 } from '@leafygreen-ui/tooltip';
 
 import { CopyButtonTrigger } from '../CodeEditorCopyButtonTrigger';
@@ -177,6 +178,7 @@ export function CodeEditorCopyButton({
       open={tooltipOpen}
       renderMode={RenderMode.TopLayer}
       setOpen={setTooltipOpen}
+      darkMode={theme === 'dark'}
       trigger={
         <CopyButtonTrigger
           variant={variant}