+++ packages/copyable/src/Copyable/Copyable.styles.ts
@@ -1,6 +1,6 @@
 import { transparentize } from 'polished';
 
-import { css } from '@leafygreen-ui/emotion';
+import { css, cx } from '@leafygreen-ui/emotion';
 import { Theme } from '@leafygreen-ui/lib';
 import { palette } from '@leafygreen-ui/palette';
 import {
@@ -13,7 +13,7 @@ import { labelTypeScaleStyles } from '@leafygreen-ui/typography';
 
 import { Size } from './Copyable.types';
 
-export const containerStyle = css`
+const containerStyle = css`
   position: relative;
   display: grid;
   grid-auto-flow: column;
@@ -24,11 +24,46 @@ export const containerStyle = css`
   margin: 2px 0;
 `;
 
-export const buttonContainerStyle = css`
+const buttonContainerStyle = css`
   height: 36px;
 `;
 
-export const codeStyle = css`
+const noButtonContainerStyle = css`
+  overflow: hidden;
+  border-radius: 12px;
+`;
+
+// When there is no button, remove the border from the code component and add the border to this component so it sits above the button wrapper box shadow
+const noButtonContainerStyleMode: Record<Theme, string> = {
+  [Theme.Light]: css`
+    border-radius: 6px;
+    border: 1px solid ${palette.gray.light2};
+  `,
+  [Theme.Dark]: css`
+    border-radius: 6px;
+    border: 1px solid ${palette.gray.dark1};
+  `,
+};
+export const getContainerStyle = ({
+  className,
+  showCopyButton,
+  theme,
+}: {
+  className?: string;
+  showCopyButton: boolean;
+  theme: Theme;
+}) =>
+  cx(
+    containerStyle,
+    {
+      [buttonContainerStyle]: showCopyButton,
+      [noButtonContainerStyleMode[theme]]: !showCopyButton,
+      [noButtonContainerStyle]: !showCopyButton,
+    },
+    className,
+  );
+
+const codeStyle = css`
   grid-area: code;
   display: inline-flex;
   align-items: center;
@@ -43,7 +78,7 @@ export const codeStyle = css`
   overflow: auto;
 `;
 
-export const codeStyleColor: Record<Theme, string> = {
+const codeStyleColor: Record<Theme, string> = {
   [Theme.Light]: css`
     color: ${palette.black};
     background-color: ${palette.gray.light3};
@@ -57,16 +92,11 @@ export const codeStyleColor: Record<Theme, string> = {
 };
 
 // Border is removed from the code component and added to the parent
-export const codeStyleNoButton = css`
+const codeStyleNoButton = css`
   border: 0;
 `;
 
-export const labelNoButtonStyle = css`
-  font-size: 18px;
-  line-height: 24px;
-`;
-
-export const codeFontStyle: Record<Size, string> = {
+const codeFontStyle: Record<Size, string> = {
   [Size.Default]: css`
     font-size: ${typeScales.code1.fontSize}px;
     line-height: ${typeScales.code1.lineHeight}px;
@@ -77,36 +107,35 @@ export const codeFontStyle: Record<Size, string> = {
   `,
 };
 
-export const labelFontStyle: Record<Size, string> = {
+const labelNoButtonStyle = css`
+  font-size: 18px;
+  line-height: 24px;
+`;
+
+const labelFontStyle: Record<Size, string> = {
   [Size.Default]: labelTypeScaleStyles[BaseFontSize.Body1],
   [Size.Large]: labelTypeScaleStyles[BaseFontSize.Body2],
 };
 
-export const noButtonContainerStyle = css`
-  overflow: hidden;
-  border-radius: 12px;
-`;
+export const getFontStyle = ({
+  size,
+  showCopyButton,
+}: {
+  size: Size;
+  showCopyButton: boolean;
+}) =>
+  cx(labelFontStyle[size], {
+    [labelNoButtonStyle]: !showCopyButton,
+  });
 
-// When there is no button, remove the border from the code component and add the border to this component so it sits above the button wrapper box shadow
-export const noButtonContainerStyleMode: Record<Theme, string> = {
-  [Theme.Light]: css`
-    border-radius: 6px;
-    border: 1px solid ${palette.gray.light2};
-  `,
-  [Theme.Dark]: css`
-    border-radius: 6px;
-    border: 1px solid ${palette.gray.dark1};
-  `,
-};
-
-export const buttonWrapperStyle = css`
+const buttonWrapperStyle = css`
   grid-area: button;
   position: relative;
   display: inline-block;
   height: 100%;
 `;
 
-export const buttonWrapperStyleShadow = css`
+const buttonWrapperStyleShadow = css`
   &::before {
     content: '';
     display: block;
@@ -120,7 +149,20 @@ export const buttonWrapperStyleShadow = css`
   }
 `;
 
-export const buttonWrapperStyleShadowTheme: Record<Theme, string> = {
+export const getCodeStyle = ({
+  theme,
+  size,
+  showCopyButton,
+}: {
+  theme: Theme;
+  size: Size;
+  showCopyButton: boolean;
+}) =>
+  cx(codeStyle, codeStyleColor[theme], codeFontStyle[size], {
+    [codeStyleNoButton]: !showCopyButton,
+  });
+
+const buttonWrapperStyleShadowTheme: Record<Theme, string> = {
   [Theme.Light]: css`
     &::before {
       box-shadow: 0 0 10px 0 ${transparentize(0.65, palette.gray.dark1)};
@@ -140,6 +182,20 @@ export const buttonWrapperStyleShadowTheme: Record<Theme, string> = {
     }
   `,
 };
+
+export const getButtonWrapperStyle = ({
+  theme,
+  isOverflowed,
+}: {
+  theme: Theme;
+  isOverflowed: boolean;
+}) =>
+  cx(buttonWrapperStyle, {
+    // Toggle these styles on only when the content extends beyond the edge of the container
+    [buttonWrapperStyleShadow]: isOverflowed,
+    [buttonWrapperStyleShadowTheme[theme]]: isOverflowed,
+  });
+
 export const buttonStyle = css`
   height: 100%;
   border-radius: 0 6px 6px 0;