+++ packages/guide-cue/src/GuideCueTooltip/GuideCueTooltip.tsx
@@ -2,13 +2,13 @@ import React from 'react';
 import { Options } from 'focus-trap';
 import FocusTrap from 'focus-trap-react';
 
-import Button from '@leafygreen-ui/button';
+import { Button } from '@leafygreen-ui/button';
 import { cx } from '@leafygreen-ui/emotion';
 import { useIdAllocator } from '@leafygreen-ui/hooks';
 import XIcon from '@leafygreen-ui/icon/dist/X';
-import IconButton from '@leafygreen-ui/icon-button';
+import { IconButton } from '@leafygreen-ui/icon-button';
 import { Theme } from '@leafygreen-ui/lib';
-import Tooltip, { RenderMode } from '@leafygreen-ui/tooltip';
+import { RenderMode, Tooltip } from '@leafygreen-ui/tooltip';
 import { Body, Disclaimer } from '@leafygreen-ui/typography';
 
 import { GuideCueProps } from '../GuideCue/GuideCue.types';