+++ packages/pipeline/src/Pipeline/Pipeline.tsx
@@ -11,7 +11,7 @@ import { cx } from '@leafygreen-ui/emotion';
 import { useMutationObserver } from '@leafygreen-ui/hooks';
 import { useDarkMode } from '@leafygreen-ui/leafygreen-provider';
 import { isComponentType } from '@leafygreen-ui/lib';
-import Tooltip, { Align, Justify, RenderMode } from '@leafygreen-ui/tooltip';
+import { Align, Justify, RenderMode, Tooltip } from '@leafygreen-ui/tooltip';
 
 import { Counter } from '../Counter';
 import { PipelineContext } from '../PipelineContext';