+++ packages/segmented-control/src/SegmentedControl/SegmentedControl.spec.tsx
@@ -3,8 +3,8 @@ import { fireEvent, getByText, render } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import { axe } from 'jest-axe';
 
-import Button from '@leafygreen-ui/button';
-import Icon from '@leafygreen-ui/icon';
+import { Button } from '@leafygreen-ui/button';
+import { Icon } from '@leafygreen-ui/icon';
 import { typeIs } from '@leafygreen-ui/lib';
 import { H1 } from '@leafygreen-ui/typography';