+++ packages/table/src/Table.Interactions.stories.tsx
@@ -6,10 +6,10 @@ import {
 import { StoryFn } from '@storybook/react';
 import { expect, userEvent, waitFor, within } from '@storybook/test';
 
-import Badge from '@leafygreen-ui/badge';
+import { Badge } from '@leafygreen-ui/badge';
 import { css } from '@leafygreen-ui/emotion';
-import Icon from '@leafygreen-ui/icon';
-import IconButton from '@leafygreen-ui/icon-button';
+import { Icon } from '@leafygreen-ui/icon';
+import { IconButton } from '@leafygreen-ui/icon-button';
 
 import { getTestUtils } from './testing/getTestUtils';
 import { makeKitchenSinkData, Person } from './utils/makeData.testutils';