+++ packages/table/src/utils/stories.testutils.tsx
@@ -1,7 +1,7 @@
 import React, { Fragment, useState } from 'react';
 import { StoryFn } from '@storybook/react';
 
-import Button from '@leafygreen-ui/button';
+import { Button } from '@leafygreen-ui/button';
 import { css, cx } from '@leafygreen-ui/emotion';
 
 import {