+++ packages/toast/src/ControlledToast/ControlledToast.stories.tsx
@@ -5,7 +5,7 @@ import {
 } from '@lg-tools/storybook-utils';
 import { StoryFn } from '@storybook/react';
 
-import Button from '@leafygreen-ui/button';
+import { Button } from '@leafygreen-ui/button';
 import { css } from '@leafygreen-ui/emotion';
 
 import { ToastProvider, Variant } from '..';