+++ packages/form-footer/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @leafygreen-ui/form-footer
 
+## 9.2.0
+
+### Minor Changes
+
+- 92693df: Updated major version to Shadow tokens. `shadow` object and its key/values have been changed. Other packages utilizing the shadow values have had a minor update to accommodate the changes.
+
+### Patch Changes
+
+- Updated dependencies [92693df]
+- Updated dependencies [c6b4d3f]
+  - @leafygreen-ui/tokens@4.0.0
+  - @leafygreen-ui/emotion@5.1.0
+  - @leafygreen-ui/icon@14.6.1
+  - @leafygreen-ui/banner@10.2.2
+  - @leafygreen-ui/button@25.1.2
+  - @leafygreen-ui/split-button@6.1.6
+
 ## 9.1.1
 
 ### Patch Changes