+++ packages/loading-indicator/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @leafygreen-ui/loading-indicator
 
+## 5.0.1
+
+### Patch Changes
+
+- Updated dependencies [92693df]
+- Updated dependencies [888a37d]
+  - @leafygreen-ui/tokens@4.0.0
+  - @leafygreen-ui/emotion@5.1.0
+  - @leafygreen-ui/typography@22.2.0
+
 ## 5.0.0
 
 ### Major Changes