+++ packages/marketing-modal/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @leafygreen-ui/marketing-modal
 
+## 8.1.3
+
+### Patch Changes
+
+- Updated dependencies [92693df]
+- Updated dependencies [888a37d]
+  - @leafygreen-ui/tokens@4.0.0
+  - @leafygreen-ui/emotion@5.1.0
+  - @leafygreen-ui/modal@20.3.0
+  - @leafygreen-ui/typography@22.2.0
+  - @leafygreen-ui/button@25.1.2
+
 ## 8.1.2
 
 ### Patch Changes