+++ packages/radio-group/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @leafygreen-ui/radio-group
 
+## 13.0.9
+
+### Patch Changes
+
+- Updated dependencies [92693df]
+- Updated dependencies [888a37d]
+  - @leafygreen-ui/tokens@4.0.0
+  - @leafygreen-ui/emotion@5.1.0
+  - @leafygreen-ui/typography@22.2.0
+  - @leafygreen-ui/hooks@9.2.2
+
 ## 13.0.8
 
 ### Patch Changes