+++ packages/segmented-control/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @leafygreen-ui/segmented-control
 
+## 11.0.10
+
+### Patch Changes
+
+- Updated dependencies [92693df]
+- Updated dependencies [c6b4d3f]
+- Updated dependencies [888a37d]
+  - @leafygreen-ui/tokens@4.0.0
+  - @leafygreen-ui/emotion@5.1.0
+  - @leafygreen-ui/icon@14.6.1
+  - @leafygreen-ui/typography@22.2.0
+  - @leafygreen-ui/hooks@9.2.2
+
 ## 11.0.9
 
 ### Patch Changes